position-service/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
pydantic==1.8.2
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
python-jose==3.3.0
jaeger-client==4.7.0
prometheus-client==0.11.0
python-consul==1.1.0
kafka-python==2.0.2
pika==1.2.0 